Fixed file issues
