+++ templates/base_model.py
@@ -1,43 +1,42 @@
-from typing import Optional, Type, Any
+import math
 
+import torch
 import torch.nn as nn
 import torch.nn.functional as F
+from torch.nn.parameter import Parameter
 
-from cogdl.trainers.base_trainer import BaseTrainer
+from .. import BaseModel, register_model
 
 
-class BaseModel(nn.Module):
+@register_model("model_name")
+class ModelName(BaseModel):
+    """
+    [Model name] [Link to paper]
+    """
+
     @staticmethod
     def add_args(parser):
-        """Add model-specific arguments to the parser."""
-        pass
+        parser.add_argument("--hidden-size", type=int, default=64)
+        parser.add_argument("--batch-size", type=int, default=20)
+        parser.add_argument("--train-ratio", type=float, default=0.7)
+        parser.add_argument("--test-ratio", type=float, default=0.1)
+        parser.add_argument("--lr", type=float, default=0.001)
 
     @classmethod
     def build_model_from_args(cls, args):
-        """Build a new model instance."""
-        raise NotImplementedError("Models must implement the build_model_from_args method")
+        return cls(args.num_features, args.hidden_size, args.num_classes,)
 
-    def __init__(self):
-        super(BaseModel, self).__init__()
-        self.device = "cpu"
+    def __init__(self, in_feats, hidden_dim, out_feats, k=20, dropout=0.5):
+        super(ModelName, self).__init__()
 
-    def _forward_unimplemented(self, *input: Any) -> None:  # abc warning
-        pass
+    def forward(self, x, adj):
+        return 0, 0  # Prediction, Loss
 
-    def forward(self, *args):
-        raise NotImplementedError
-
-    def node_classification_loss(self, data):
-        pred = self.forward(data.x, data.edge_index)
-        pred = F.log_softmax(pred, dim=-1)
+    def loss(self, data):
         return F.nll_loss(
-            pred[data.train_mask],
+            self.forward(data.x, data.edge_index)[data.train_mask],
             data.y[data.train_mask],
         )
 
-    @staticmethod
-    def get_trainer(taskType: Any, args: Any) -> Optional[Type[BaseTrainer]]:
-        return None
-
-    def set_device(self, device):
-        self.device = device
+    def predict(self, data):
+        return self.forward(data.x, data.edge_index)
